\name{exploBATCHbreast-package}
\alias{exploBATCHbreast-package}
\alias{exploBATCHbreast}
\docType{package}
\title{
\packageTitle{exploBATCHbreast}
}
\description{
\packageDescription{exploBATCHbreast}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{exploBATCHbreast}
\packageIndices{exploBATCHbreast}
A package with breast cancer gene expression dataset used in Nyamundanda et al (2017) manuscript on batch detection and correction.
}
\author{
\packageAuthor{exploBATCHbreast}

Maintainer: \packageMaintainer{exploBATCHbreast}
}
\references{
Nyamundanda, G., Poudel, P., Patil, Y. and Sadanandam, A., 2017. A novel and robust statistical method to diagnose and correct batch effects in genomic data.
}
\examples{
data(Breast)  # Data for the first example in the manuscript.
}
